Add abstract boxes and glue algorithm; not everything is implemented but it should already be pretty useful
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5689 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/bignum/big.h
@@ -1083,7 +1083,7 @@ double BigInt::ToDouble() const {
     uint64_t udigit = BzGetDigit(rep, idx);
     double ddigit = udigit;
     // printf("d %.17g | u %llu | dd %.17g\n", d, udigit, ddigit);
-    for (int e = 0; e < sizeof (BigNumDigit); e++)
+    for (size_t e = 0; e < sizeof (BigNumDigit); e++)
       d *= 256.0;
     d += ddigit;
   }
